Fixed key press and releases for unity games
Code is still buggy for two keys in one packet. E.g. UP RIGHT.

This is due to the fact that on the server side, each keypress delays
for 100ms so two keys = 200ms, but the client side sends in packets
every 100ms.

Recommended fix is to change server side so that in the event of 2 keys,
the delay is 50ms each rather than 100ms